+++ tests/Header/AuthorizationValueTest.php
@@ -19,7 +19,7 @@ class AuthorizationValueTest extends TestCase
         $this->assertInstanceOf(Value::class, $a);
         $this->assertSame('Basic', $a->scheme());
         $this->assertSame('realm', $a->parameter());
-        $this->assertSame('"Basic" realm', $a->toString());
+        $this->assertSame('Basic realm', $a->toString());
 
         new AuthorizationValue('Basic', '');
         new AuthorizationValue('Basic', 'realm="some value"');
@@ -43,6 +43,7 @@ class AuthorizationValueTest extends TestCase
             ['foo@bar'],
             ['foo/bar'],
             ['"Basic"realm'],
+            ['"Basic" realm'],
         ];
     }
 }